The so needed IE's counter example
In order to throw down IE there was a need of a counter example.
While studing many hipotesis I came up with a possible counter example.
These changes are the apply of that counter example that shows that IE's defer
has actually a fault like many said.

This does not mean we cannot use the defer completly with IE, it means that
if we are including defered scripts from an external domain (scripts from
another domain) scripts into IE then the external scripts are executed
sometime in the future.