The code to test browsers
This code is made to test if browsers follow the defer rules.
Known results:
Firefox > 3.6.10 (gecko), Chrome (webkit), IE 10 (IE!!!!) passed 15/15
IE < 10 has issues about defer forcing the download of jQuery (differnt domain)\
to be synced but the rest can use the defer without any problems.
Opera passed 9/15 not because it has a messed up system but because it ignores\
defer.
Any complaint is welcome to be discussed.